Deposit fetcher added

+++ FirstSparrow.Application/Domain/Entities/Deposit.cs
@@ -6,7 +6,7 @@ public class Deposit : BaseEntity<int>
 {
     public string Commitment { get; set; }
 
-    public ulong  LeafIndex { get; set; }
+    public uint  LeafIndex { get; set; }
 
     public DateTime DepositTimestamp { get; set; }
 }
